% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assemblage_nspecies.R
\name{assemblage_nspecies}
\alias{assemblage_nspecies}
\title{Calculate the number of species in each grid/assemblage/site}
\usage{
assemblage_nspecies(
  df.TS.TE,
  df.occ,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  site = "site"
)
}
\arguments{
\item{site}{}
}
\description{
Calculate the number of species in each grid/assemblage/site
}
