% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assemblage_regional_mpd.R
\name{Assemblage_regional_mpd}
\alias{Assemblage_regional_mpd}
\title{Compute mean pairwise distance for each site in time slices}
\usage{
Assemblage_regional_mpd(
  df.TS.TE,
  df.occ,
  time.slice,
  grid.size,
  trait,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  lat = "lat",
  lon = "lng",
  crs = 4326
)
}
\arguments{
\item{crs}{}
}
\description{
Compute mean pairwise distance for each site in time slices
}
