% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assemblage_regional_richness.R
\name{Assemblage_regional_richness}
\alias{Assemblage_regional_richness}
\title{Mean richness by communities in a time series}
\usage{
Assemblage_regional_richness(
  df.TS.TE,
  df.occ,
  time.slice,
  grid.size,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  lat = "lat",
  lon = "lng",
  crs = 4326
)
}
\arguments{
\item{df.TS.TE}{Data frame object containing at least four columns. Species names,
origination time, extinction time and a trait value for each species.}

\item{df.occ}{a data frame object containing the occurrence records for each species.
This must have at least a column indicating the name of species, its minimum and maximum age estimate,
and its site location ID.}

\item{time.slice}{Scalar indicating the time interval between consecutive time slices.}

\item{grid.size}{Scalar indicating the grid size in degrees.}

\item{round.digits}{Scalar indicating the precision of time slices.}

\item{species}{Character indicating the name of the column of the data frame
containing the species name information.}

\item{TS}{Character indicating the name of the columns of the data frame
containing the information on origination time.}

\item{TE}{Character indicating the name of the column of the data frame
containing the information on extinction time.}

\item{Max.age}{Character indicating the name of the column containing the upper age limit for occurrence record.}

\item{Min.age}{Character indicating the name of the column containing the lower age limit for occurrence record.}

\item{lat}{Character indicating the name of the column containing the latitude
information for occurrence record. Default is "lat"}

\item{lon}{Character indicating the name of the column containing the latitude
information for occurrence record. Default is "lng"}

\item{crs}{A scalar indicating the projection used. Default is 4326}
}
\description{
Mean richness by communities in a time series
}
