% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_matrix_regional_coex.R
\name{aux_matrix_regional_coex}
\alias{aux_matrix_regional_coex}
\title{Auxiliary function to compute regional coexistence matrix}
\usage{
aux_matrix_regional_coex(
  df.TS.TE,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{}

\item{time.slice}{}

\item{round.digits}{}

\item{species}{}

\item{TS}{}

\item{TE}{}
}
\description{
Auxiliary function to compute regional coexistence matrix
}
