% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_mpd_general.R
\name{aux_mpd_general}
\alias{aux_mpd_general}
\title{Auxiliar function to calculate trait distance considering different distance thresholds}
\usage{
aux_mpd_general(df.TS.TE, dist.trait, nearest.taxon)
}
\arguments{
\item{nearest.taxon}{}
}
\description{
Auxiliar function to calculate trait distance considering different distance thresholds
}
