% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_site_cooccurrence.R
\name{calc_cooccur_site_matrix}
\alias{calc_cooccur_site_matrix}
\title{Calculate site cooccurrence}
\usage{
calc_cooccur_site_matrix(spp_slice, df.occ)
}
\arguments{
\item{list_site_interval}{}
}
\description{
Calculate site cooccurrence
}
