% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_function_null_model_sites_grids.R
\name{calc_null_model}
\alias{calc_null_model}
\title{Compute null model with fixed richness for sites and assemblage}
\usage{
calc_null_model(
  list_occurrence,
  dist_matrix_trait,
  nearest.taxon,
  nperm = 1000
)
}
\arguments{
\item{list_occurrence}{a data frame containing species occurrence in sites
or assemblages/grids}

\item{dist_matrix_trait}{a distance matrix computed}

\item{nperm}{the number of permutations for the null model}
}
\value{
a data frame with distance metric for all time slices
}
\description{
Compute null model with fixed richness for sites and assemblage
}
