% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-reach-distance2.R
\name{calc_reach}
\alias{calc_reach}
\title{Auxiliar function to calculate reach criteria}
\usage{
calc_reach(df.coords)
}
\arguments{
\item{df.coords}{}
}
\description{
Auxiliar function to calculate reach criteria
}
