% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_calc_spp_slice.R
\name{calc_spp_slice}
\alias{calc_spp_slice}
\title{Auxiliar function to compute the names of species in each timeslice}
\usage{
calc_spp_slice(
  df.TS.TE,
  time.slice,
  seq.interval,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{TE}{}
}
\value{
a list with names of species in each timeslice
}
\description{
Auxiliar function to compute the names of species in each timeslice
}
