% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_regional_richness.R
\name{clade_regional_richness}
\alias{clade_regional_richness}
\title{Calculates regional clade richness for lineages}
\usage{
clade_regional_richness(
  df.TS.TE,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{Data frame object containing at least four columns. Species names,
origination time, extinction time and a trait value for each species.}

\item{time.slice}{Scalar indicating the time interval between consecutive time slices.}

\item{round.digits}{Scalar indicating the precision of time slices.}

\item{species}{Character indicating the name of the column of the data frame
containing the species name information.}

\item{TS}{Character indicating the name of the columns of the data frame
containing the information on origination time.}

\item{TE}{Character indicating the name of the column of the data frame
containing the information on extinction time.}
}
\value{
A data frame with richness information in each timeslice
}
\description{
Calculates regional clade richness for lineages
}
