% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_site_richness.R
\name{clade_site_richness}
\alias{clade_site_richness}
\title{Compute time series based on mean site richness}
\usage{
clade_site_richness(
  df.TS.TE,
  df.occ,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{}

\item{df.occ}{}

\item{time.slice}{}

\item{round.digits}{}

\item{species}{}

\item{TS}{}

\item{TE}{}
}
\description{
Compute time series based on mean site richness
}
