% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_longevities_canidae}
\alias{df_longevities_canidae}
\title{Longevities for 133 Canidae species}
\format{
A data frame object with 133 rows and 3 columns:
\describe{
\item{species}{Species names}
\item{TS}{Time of Speciation}
\item{TE}{Time of Extinction}
}
}
\source{
\url{https://academic.oup.com/evolut/article/79/9/1835/8140865#535027902}
}
\usage{
df_longevities_canidae
}
\description{
Data frame containing Time of Speciation (TS) and Time of Extinction (TE) for
133 Canidae species
}
\keyword{datasets}
