% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_occurrence_canidae}
\alias{df_occurrence_canidae}
\title{Occurrence records}
\format{
A data frame object with 133 rows and 14 columns:
\describe{
\item{species}{Species name}
\item{MaxT}{Maximum (oldest) age estimate for each occurrence record}
\item{MinT}{the minimum (youngest) age estimate for each occurrence record}
\item{lng}{The longitude coordinate of the occurrence record}
\item{lat}{The latitude coordinate of the occurrence record}
\item{Site}{Name of the paleontological site location identifier}
\item{max_low_res}{Same as MaxT but with lower temporal precision}
\item{min_lo_res}{Same as MinT but with lower temporal precision}
\item{site.char}{Name of the paleontological site identifier}
\item{midpoint}{Age midpoint of occurrence record}
}
}
\source{
\url{https://academic.oup.com/evolut/article/79/9/1835/8140865#535027902}
}
\usage{
df_occurrence_canidae
}
\description{
Data frame containing spatial occurrence records of 133 species of canidae
in North America
}
\keyword{datasets}
