% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_interval.R
\name{get_summary_interval}
\alias{get_summary_interval}
\title{Compute summary statistics from fossil record}
\usage{
get_summary_interval(
  df.occ.fossil,
  interval,
  age.occ = "midpoint",
  species = "species",
  Max.age = "Maximum_Age",
  Min.age = "Minimum_Age",
  TS = NULL,
  TE = NULL,
  lat = NULL,
  lng = NULL,
  site = NULL,
  group = NULL,
  trait = NULL,
  ...
)
}
\arguments{
\item{df.occ.fossil}{}

\item{interval}{}

\item{age.occ}{}

\item{species}{}

\item{Max.age}{}

\item{Min.age}{}

\item{TS}{}

\item{TE}{}

\item{lat}{}

\item{lng}{}

\item{site}{}

\item{group}{}

\item{trait}{}

\item{...}{}
}
\description{
Compute summary statistics from fossil record
}
