% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_slice.R
\name{get_summary_slices}
\alias{get_summary_slices}
\title{Getting summary information for time bins}
\usage{
get_summary_slices(
  df.TS.TE,
  timeframe,
  method.ages = c("midpoint", "upper", "lower"),
  thresh.age.range = 10,
  species = "species",
  Max.age = "Maximum_Age",
  Min.age = "Minimum_Age",
  TS = NULL,
  TE = NULL,
  lat = NULL,
  lng = NULL,
  site = NULL,
  group = NULL,
  trait = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Getting summary information for time bins
}
