% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivSpecies_reach_richness.R
\name{IndivSpec_reach_richness}
\alias{IndivSpec_reach_richness}
\title{Compute individual species coexistence based on reach criteria}
\usage{
IndivSpec_reach_richness(
  df.TS.TE,
  df.occ,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  lat = "lat",
  lon = "lng"
)
}
\arguments{
\item{df.TS.TE}{a data frame object containing at least three columns. Species names,
origination time and extinction time for each species.}

\item{df.occ}{a data frame object containing the occurrence records for each species.
This must have at least a column indicating the name of species, its minimum and maximum age estimate,
and its site location ID.}

\item{time.slice}{scalar indicating the time interval between consecutive time slices.}

\item{round.digits}{scalar indicating the number of digits for time of origination and time for
extinction.}

\item{species}{character indicating the name of the column of the data frame
containing the species name information.}

\item{TS}{character indicating the name of the columns of the data frame
containing the information on origination time.}

\item{TE}{character indicating the name of the column of the data frame
containing the information on extinction time.}

\item{Max.age}{character indicating the name of the column containing the upper age limit for occurrence record.}

\item{Min.age}{character indicating the name of the column containing the lower age limit for occurrence record.}

\item{lat}{character indicating the name of the column containing the latitude of occurrence record in df.occ.}

\item{lon}{character indicating the name of the column containing the longitude of occurrence record in df.occ.}
}
\value{
data frame containing the name of species, its mean coexistence value calculated by each time slice
considering reach criteria of co-occurrence
}
\description{
Compute individual species coexistence based on reach criteria
}
