% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivSpecies_site_mpd.R
\name{IndivSpec_site_mpd}
\alias{IndivSpec_site_mpd}
\title{Compute mean pairwise distance for each species co-occurring in sites and time slices}
\usage{
IndivSpec_site_mpd(
  df.TS.TE,
  df.occ,
  time.slice,
  trait,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  site = "site"
)
}
\arguments{
\item{df.TS.TE}{a data frame object containing at least three columns. Species names,
origination time and extinction time for each species.}

\item{df.occ}{a data frame object containing the occurrence records for each species.
This must have at least a column indicating the name of species, its minimum and maximum age estimate,
and its site location ID.}

\item{time.slice}{scalar indicating the time interval between consecutive time slices.}

\item{trait}{Numeric indicating the values of the traits for each species}

\item{round.digits}{scalar indicating the number of digits for time of origination and time for
extinction.}

\item{species}{character indicating the name of the column of the data frame
containing the species name information.}

\item{TS}{character indicating the name of the columns of the data frame
containing the information on origination time.}

\item{TE}{character indicating the name of the column of the data frame
containing the information on extinction time.}

\item{Max.age}{character indicating the name of the column in df.occ containing the upper age limit for occurrence record.}

\item{Min.age}{character indicating the name of the column in df.occ containing the lower age limit for occurrence record.}

\item{site}{character indicating the name of the column in df.occ containing the information on site location.}
}
\value{
A data frame containing the name of species, its mean mpd value calculated by each time slice
considering all sites in which the species occur in that time slice
}
\description{
Compute mean pairwise distance for each species co-occurring in sites and time slices
}
