% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivSpecies_regional_mpd.R
\name{IndivSpecies_regional_mpd}
\alias{IndivSpecies_regional_mpd}
\title{Compute regional time series of Individual Species Mean Pairwise trait distance}
\usage{
IndivSpecies_regional_mpd(
  df.TS.TE,
  time.slice,
  trait = NULL,
  dist.trait = NULL,
  nearest.taxon = "all",
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{a data frame object containing at least four columns. Species names,
origination time, extinction time and a trait value for each species.}

\item{time.slice}{scalar indicating the time interval between consecutive time slices}

\item{trait}{Character indicating the name of the column containing values of the traits for each species. If NULL,
the default, the user must provide a distance matrix}

\item{dist.trait}{A distance matrix containing pairwise distance between species. If NULL
the default, the user must provide a trait in trait argument}

\item{nearest.taxon}{scalar indicating the number of closest species that will be used to compute
the mean pairwise distance. If "all", the default, all species coexisting in a time slice
will be used to calculate mpd}

\item{round.digits}{scalar indicating the number of digits for time of origination and time for
extinction}

\item{species}{character indicating the name of the column of the data frame
containing the species name information}

\item{TS}{character indicating the name of the columns of the data frame
containing the information on origination time}

\item{TE}{character indicating the name of the column of the data frame
containing the information on extinction time}
}
\value{
A data frame containing the name of species and the mean value of mpd of each species
in each time slice
}
\description{
Compute regional time series of Individual Species Mean Pairwise trait distance
}
