% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndivSpecies_regional_richness.R
\name{IndivSpecies_regional_richness}
\alias{IndivSpecies_regional_richness}
\title{Calculate individual species coexistence in regional scale}
\usage{
IndivSpecies_regional_richness(
  df.TS.TE,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{a data frame object containing at least three columns. Species names,
origination time and extinction time for each species.}

\item{time.slice}{scalar indicating the time interval between consecutive time slices}

\item{round.digits}{scalar indicating the number of digits for time of origination and time for
extinction}

\item{species}{character indicating the name of the column of the data frame
containing the species name information}

\item{TS}{character indicating the name of the columns of the data frame
containing the information on origination time}

\item{TE}{character indicating the name of the column of the data frame
containing the information on extinction time}
}
\value{
A list with two elements:
\item{df_IndivSpp_coexist}{A data frame with species name, the number of coexistence for each sepecies in each time slice}
\item{mean_species_coexistence}{A data frame with three column. The species name, the mean number of coexistence for each species
and the variance considering all time slices}
}
\description{
Calculate individual species coexistence in regional scale
}
