% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_range.R
\name{plot_occ_bins}
\alias{plot_occ_bins}
\title{Plot occurrence records and age ranges}
\usage{
plot_occ_bins(
  df.occ.fossil,
  species = "accepted_name",
  max.age = "max_ma",
  min.age = "min_ma",
  occ.gradient = TRUE,
  log = TRUE,
  name.scale.occ = "viridis",
  x.axis.name = "age range",
  y.axis.name = "species.name",
  show.species.name = FALSE,
  age.scheme = TRUE
)
}
\arguments{
\item{df.occ.fossil}{}

\item{species}{}

\item{max.age}{}

\item{min.age}{}

\item{occ.gradient}{}

\item{log}{}

\item{name.scale.occ}{}

\item{x.axis.name}{}

\item{y.axis.name}{}

\item{show.species.name}{}
}
\description{
Plot occurrence records and age ranges
}
