% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_regional_richness.R
\name{regional_clade_richness}
\alias{regional_clade_richness}
\title{Calculates regional clade richness for lineages}
\usage{
regional_clade_richness(
  df.TS.TE,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{}

\item{time.slice}{}

\item{round.digits}{}

\item{species}{}

\item{TS}{}

\item{TE}{}
}
\description{
Calculates regional clade richness for lineages
}
