% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_ses_mpd_modif.R
\name{ses.mpd.modif}
\alias{ses.mpd.modif}
\title{Title}
\usage{
ses.mpd.modif(
  samp,
  dis,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool",
    "independentswap", "trialswap"),
  abundance.weighted = FALSE,
  runs = 999,
  iterations = 1000
)
}
\arguments{
\item{iterations}{}
}
\description{
Title
}
