% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits_canidae}
\alias{traits_canidae}
\title{Trait data for 133 Canidae species}
\format{
A data frame object with 133 rows and 5 columns, each row
correspond to a species name their respective traits:
\describe{
\item{species}{Species name}
\item{diet}{Numeric indicating the diet category of species}
\item{LD1}{Numeric indicating the level of hypercarnivore of species}
\item{log_mass}{Numeric indicating the mass of species in log scale}
\item{diet_cat}{Character indicating the diet category of species.
hyper indicate hypercarnivores, meso indicates mesocarnivores and
hypo hypocarnivores}
}
}
\source{
\url{https://academic.oup.com/evolut/article/79/9/1835/8140865#535027902}
}
\usage{
traits_canidae
}
\description{
Data frame containing five traits measured for 133 Canidae species
}
\keyword{datasets}
