% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assemblage_nspecies.R
\name{assemblage_nspecies}
\alias{assemblage_nspecies}
\title{Calculate the number of species in each grid/assemblage/site}
\usage{
assemblage_nspecies(
  df.TS.TE,
  df.occ,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  site = "site"
)
}
\arguments{
\item{df.TS.TE}{Data frame object containing at least four columns. Species names,
origination time, extinction time and a trait value for each species.}

\item{df.occ}{Data frame object containing the occurrence records for each species.
This must have at least a column indicating the name of species, its minimum and maximum age estimate,
and its site location ID.}

\item{time.slice}{Scalar indicating the time interval between consecutive time slices.}

\item{round.digits}{Scalar indicating the precision of time slices.}

\item{species}{Character indicating the name of the column of the data frame
containing the species name information.}

\item{TS}{Character indicating the name of the columns of the data frame
containing the information on origination time.}

\item{TE}{Character indicating the name of the column of the data frame
containing the information on extinction time.}

\item{Max.age}{Character indicating the name of the column containing the upper
age limit for occurrence record.}

\item{Min.age}{Character indicating the name of the column containing the
lower age limit for occurrence record.}

\item{site}{Character indicating the name of the column containing the
information on assemblage location.}
}
\description{
Calculate the number of species in each grid/assemblage/site
}
