% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assemblage_regional_richness.R
\name{Assemblage_regional_richness}
\alias{Assemblage_regional_richness}
\title{Mean richness by communities in a time series}
\usage{
Assemblage_regional_richness(
  df.TS.TE,
  df.occ,
  time.slice,
  grid.size,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  Max.age = "Max.age",
  Min.age = "Min.age",
  lat = "lat",
  lon = "lng",
  crs = 4326
)
}
\arguments{
\item{df.TS.TE}{}

\item{df.occ}{}

\item{time.slice}{}

\item{grid.size}{}

\item{round.digits}{}

\item{species}{}

\item{TS}{}

\item{TE}{}

\item{Max.age}{}

\item{Min.age}{}

\item{lat}{}

\item{lon}{}

\item{crs}{}
}
\description{
Mean richness by communities in a time series
}
