% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crossing_boundaries.R
\name{check_crossing_boundaries}
\alias{check_crossing_boundaries}
\title{Check occurrence records with crossing boundaries in time bins}
\usage{
check_crossing_boundaries(
  df.occ.fossil,
  interval,
  species = "species",
  Max.age = "Maximum_Age",
  Min.age = "Minimum_Age",
  remove.sub.species = TRUE
)
}
\arguments{
\item{df.occ.fossil}{}

\item{interval}{}

\item{species}{}

\item{Max.age}{}

\item{Min.age}{}

\item{remove.sub.species}{}
}
\description{
Check occurrence records with crossing boundaries in time bins
}
