% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_regional_coexistence.R
\name{clade_regional_coexistence}
\alias{clade_regional_coexistence}
\title{Calculates regional clade coexistence for lineages}
\usage{
clade_regional_coexistence(
  df.TS.TE,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)
}
\arguments{
\item{df.TS.TE}{Data frame object containing at least three columns. Species names,
origination time, extinction time.}

\item{time.slice}{Scalar indicating the time interval between consecutive
time slices in which coexistence will be calculated.}

\item{round.digits}{Scalar indicating the precision of time slices. This corresponds
to the number of digits to be considered in the calculation of time
coexistence in each time slice.}

\item{species}{Character indicating the name of the column of the data frame
containing the species name information. The default is "species".}

\item{TS}{Character indicating the name of the columns of the data frame
containing the information on origination time. The default is "TS".}

\item{TE}{Character indicating the name of the column of the data frame
containing the information on extinction time. The default is "TE".}
}
\value{
A two column data frame object containing the number of coexisting
clades (coexistence) at each time slice (time.slice).
}
\description{
Calculates regional clade coexistence for lineages
}
\examples{
data(longevities_canidae) # list containing longevities for canidae species
longs_1 <- data.frame(longs[[1]], species = rownames(longs[[1]])) # extracting only one replicate
res_regional_coexistence <- clade_regional_coexistence(df.TS.TE = longs_1,
                                                       time.slice,
                                                       round.digits = 1,
                                                       species = "species",
                                                       TS = "TS",
                                                       TE = "TE")


}
