% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_regional_trait_distance.R
\name{clade_regional_distance}
\alias{clade_regional_distance}
\title{Compute regional clade Mean Trait Distance in a time series}
\usage{
clade_regional_distance(
  df.TS.TE,
  time.slice,
  dist.trait,
  nearest.taxon,
  trait = NULL,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  group = NULL,
  group.focal.compare = NULL,
  type.comparison = NULL
)
}
\arguments{
\item{df.TS.TE}{Data frame object containing at least four columns. Species names,
origination time, extinction time and a trait value for each species.}

\item{time.slice}{Scalar indicating the time interval between consecutive time slices.}

\item{dist.trait}{A dist object containing the pairwise distance matrices}

\item{nearest.taxon}{A scalar indicating the number of nearest species/genus that will be used.
1 computes mnnd metric and the option "all" computes mpd.}

\item{trait}{Character indicating the name of the column containing values of the traits for each species. If NULL,
the default, the user must provide a distance matrix.}

\item{round.digits}{Scalar indicating the precision of time slices.}

\item{species}{Character indicating the name of the column of the data frame
containing the species name information.}

\item{TS}{Character indicating the name of the columns of the data frame
containing the information on origination time.}

\item{TE}{Character indicating the name of the column of the data frame
containing the information on extinction time.}

\item{group}{Character indicating the name of the column that contain the groups that will be used
in comparison.}

\item{group.focal.compare}{Character vector indicating the focal (first element) and comparison (second element).}

\item{type.comparison}{Character. It can be "between" to compute distances only between species/genus of two groups
or "within" to calculate distance only inside the focal group.}
}
\description{
Compute regional clade Mean Trait Distance in a time series
}
