% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_site_matrix_coocurrence.R
\name{comp_site_cooccurr}
\alias{comp_site_cooccurr}
\title{Auxiliary function to compute site cooccurrence matrix per timeslice}
\usage{
comp_site_cooccurr(spp_slice, df.occ)
}
\arguments{
\item{spp_slice}{A list with species name in each timeslice}

\item{df.occ}{a occurrence data frame with the name of species and the sites in which
they occur}
}
\description{
Auxiliary function to compute site cooccurrence matrix per timeslice
}
