% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_function_site_grid_composition.R
\name{comp_site_occurrence}
\alias{comp_site_occurrence}
\title{Auxiliar function to compute site and grid composition}
\usage{
comp_site_occurrence(
  spp_slice,
  df.occ,
  species = "species",
  Max.age = "Max.age",
  Min.age = "Min.age",
  site = "site"
)
}
\arguments{
\item{site}{}
}
\description{
Auxiliar function to compute site and grid composition
}
